requirements.txt
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Django==2.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.8
django-extensions==2.1.6
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
httplib2==0.12.3
idna==2.8
oauthlib==3.0.1
Pillow==6.0.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.3
psycopg2==2.7.7
simpleduration
